contracts(feat): ligitateContent adds missing content
