+++ test/litigate.ts
@@ -1,38 +1,40 @@
 import { expect } from "chai";
-import { ethers } from "hardhat";
-import { deployContractFixture } from "./fixtures";
+import { prepareEIP712LitigateContentFixture } from "./fixtures";
 import { loadFixture } from "@nomicfoundation/hardhat-network-helpers";
-import { BACKEND_PRIVATE_KEY } from "../constants";
+import { FIRST_TAG, SECOND_TAG } from "../constants";
 
 describe("Litigate", async function () {
-    it("Should succesfully litigate a content", async function () {
-        const { channel4Contract, otherAccount1 } = await loadFixture(deployContractFixture);
-        const EIP712Domain = await channel4Contract.eip712Domain();
-        const domain = {
-            name: EIP712Domain.name,
-            version: EIP712Domain.version,
-            chainId: EIP712Domain.chainId,
-            verifyingContract: EIP712Domain.verifyingContract
-        };
-        const types = {
-            ContentToAdd: [
-                { name: 'title', type: 'string' },
-                { name: 'url', type: 'string' },
-                { name: 'submittedBy', type: 'address' },
-                { name: 'likes', type: 'uint256' },
-                { name: 'tagIds', type: 'string[]' },
-            ],
-        };
+    it("Should add a missing content", async function () {
+        const { channel4Contract, otherAccount1, domain, types, backendWallet } = await loadFixture(prepareEIP712LitigateContentFixture);
         const content = {
             title: 'Batman',
             url: 'https://www.dc.com/characters/batman',
             submittedBy: otherAccount1.address,
             likes: 0,
-            tagIds: ['zero-knowledge'],
+            tagIds: ['superhero'],
         };
-        const backendWallet = new ethers.Wallet(BACKEND_PRIVATE_KEY);
         const EIPSignature = await backendWallet.signTypedData(domain, types, content);
         const result = await channel4Contract.litigateContent(content, EIPSignature);
+
+        const allContent = await channel4Contract.getAllContent();
+        console.log(allContent);
+
+        console.log(result);
+        // TODO: setup expects
+    });
+
+    it("Should prevent litigation in correct content", async function () {
+        const { channel4Contract, otherAccount1, domain, types, backendWallet } = await loadFixture(prepareEIP712LitigateContentFixture);
+        const content = {
+            title: 'Google',
+            url: 'https://google.com/',
+            submittedBy: otherAccount1.address,
+            likes: 0,
+            tagIds: [FIRST_TAG, SECOND_TAG],
+        };
+        const EIPSignature = await backendWallet.signTypedData(domain, types, content);
+        const result = await channel4Contract.litigateContent(content, EIPSignature);
+        console.log(result);
         // TODO: setup expects
     });
 });
